requirements.txt
tabulate
colorama
requests
python-dotenv